backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation
pydantic==2.5.0

# OpenRouter/HTTP
httpx==0.25.0
openai==1.3.0

# Environment
python-dotenv==1.0.0

# Database
sqlite3  # Built into Python

# Async
asyncio  # Built into Python

# For Hugging Face Spaces deployment
gradio==4.8.0
